feat: migrate archimate to module convention
- MODULE.yaml, v1/tools.py (8 tools), v1/module.py, router server.py
- Dockerfile updated to COPY . .
- docker-compose: mcp-archimate → module-archimate
- Volume mount path updated for models directory
