+++ druppie/mcp-servers/module-archimate/MODULE.yaml
@@ -1,0 +1,4 @@
+id: archimate
+latest_version: "1.0.0"
+versions:
+  - "1.0.0"
