+++ druppie/mcp-servers/module-archimate/requirements.txt
@@ -1,2 +1,3 @@
 fastmcp>=2.0.0,<3.0.0
 uvicorn
+pyyaml
