+++ druppie/mcp-servers/module-hitl/MODULE.yaml
@@ -1,0 +1,4 @@
+id: hitl
+latest_version: "1.0.0"
+versions:
+  - "1.0.0"
