+++ druppie/mcp-servers/module-hitl/requirements.txt
@@ -1,3 +1,4 @@
 fastmcp>=2.0.0,<3.0.0
 httpx>=0.25.0
 uvicorn>=0.24.0
+pyyaml
